move as_contiguous_vec to utils, use it also for node masks
