+++ meshkernel/c_structures.py
@@ -19,25 +19,7 @@ from meshkernel.py_structures import (
     OrthogonalizationParameters,
     SplinesToCurvilinearParameters,
 )
-
-
-def as_contiguous_vec(vec) -> np.ndarray:
-    """Ensures the input vector is contiguous before passing it to a MeshKernel C API function,
-     if the vector has been created with slicing operations.
-
-    Args:
-        vec (np.ndarray): The input vector.
-
-    Returns:
-        np.ndarray: The contiguous vector.
-
-    Raises:
-        TypeError: If `vec` is not a NumPy array.
-    """
-    if not isinstance(vec, np.ndarray):
-        raise TypeError("Input must be a NumPy array in as_contiguous_vec function.")
-
-    return np.ascontiguousarray(vec)
+from meshkernel.utils import as_contiguous_vec
 
 
 class CMesh2d(Structure):
