+++ meshkernel/utils.py
@@ -4,6 +4,25 @@ import numpy as np
 from matplotlib.collections import LineCollection
 
 
+def as_contiguous_vec(vec) -> np.ndarray:
+    """Ensures the input vector is contiguous before passing it to a MeshKernel C API function,
+     if the vector has been created with slicing operations.
+
+    Args:
+        vec (np.ndarray): The input vector.
+
+    Returns:
+        np.ndarray: The contiguous vector.
+
+    Raises:
+        TypeError: If `vec` is not a NumPy array.
+    """
+    if not isinstance(vec, np.ndarray):
+        raise TypeError("Input must be a NumPy array in as_contiguous_vec function.")
+
+    return np.ascontiguousarray(vec)
+
+
 def plot_edges(node_x, node_y, edge_nodes, ax, *args, **kwargs):
     """Plots the edges at a given axes.
     `args` and `kwargs` will be used as parameters of the `plot` method.
